Добавила удаление товара из корзины, вывод моих заказов и их содержимого, подправила редактирование профиля и добавление товара из каталога(чтобы не добавлял больше, чем есть на складе), добавила редактирование и удаление отзыва,

+++ StoreTeddyBear/StoreTeddyBear/Controllers/ReviewController.cs
@@ -80,7 +80,7 @@ namespace StoreTeddyBear.Controllers
         }
 
         [HttpPut("{reviewId}/Edit")]
-        public ActionResult<Review> EditReview(int reviewId, int rating, string comment)
+        public ActionResult<Review> EditReview(int reviewId, sbyte rating, string comment)
         {
             var review = StorepinkteddybearBdContext.Instance.Reviews
                 .Find(reviewId);
